third_party/WebKit/LayoutTests/external/wpt/css/css-animations/CSSAnimation-playState.tentative-expected.txt

This is a testharness.js-based test.
FAIL A new CSS animation is initially play-pending assert_true: expected true got undefined
FAIL Animation returns correct playState when paused assert_equals: expected "paused" but got "pending"
FAIL Animation.playState updates when paused by script assert_equals: expected "paused" but got "pending"
FAIL Animation.playState updates when resumed by setting style assert_equals: Animation.playState reports running after updating animation-play-state (got: pending) expected "running" but got "pending"
PASS Animation returns correct playState when canceled
Harness: the test ran to completion.
